Versao estavel. Interligacao entre todas as classes.

+++ src/Arduino/Leg.cpp
@@ -1,10 +1,11 @@
 #include <Servo.h>
 #include <Arduino.h>
 #include "Leg.h"
+#include <stdint.h>
 
 
 //inicializador de perna. 
-Leg::Leg(int coxapos, int femurpos, int tibiapos, int initcoxa, int initfemur, int inittibia){
+Leg::Leg(uint8_t coxapos, uint8_t femurpos, uint8_t tibiapos, uint8_t initcoxa, uint8_t initfemur, uint8_t inittibia){
 	coxa.attach(initcoxa);
 	femur.attach(initfemur);
 	tibia.attach(inittibia);
@@ -17,7 +18,7 @@ Leg::~Leg(){
 }
 
 //posiciona os 3 servos da perna nas posicoes indicadas por posCoxa, posFemur e posTibia
-void Leg::move(int posCoxa, int posFemur, int posTibia){
+void Leg::move(uint8_t posCoxa, uint8_t posFemur, uint8_t posTibia){
 	coxa.write(posCoxa);
 	femur.write(posFemur);
 	tibia.write(posTibia);
